+++ src/main/java/org/y1000/entities/players/kungfu/LevelKungFu.java
@@ -2,6 +2,6 @@ package org.y1000.entities.players.kungfu;
 
 public interface LevelKungFu extends KungFu {
 
-    float level();
+    int level();
 
 }
